fix compiler warnings (#19221)
* fix compiler warnings

* allow hashtable values to be non-pointers; make hashtable use 64bit values even on 32bit machines

* make sure hashtable always returns the right type of value

* simple_hashtable_set_slot_named() fixes

* fix warning in simple hashtable

* more fixes

* have a custom preprocessor directive for non-pointers seems the only viable solution

* uint64_t is %llX on 32bit

* uint64_t is %llX on 32bit and it needs casting on 64bit

* always print 16 digits

* fix windows hashtable

* fix more warnings

* another

* fixed warnings on windows

* do not show pulse dbengine charts if extended mode is not enabled

* fix more warnings again

+++ src/streaming/stream-compression/compression.c
@@ -576,8 +576,8 @@ cleanup:
         fprintf(stderr, "Compression with %s: FAILED (%d errors)\n", name, errors);
     else
         fprintf(stderr, "Compression with %s: OK "
-                        "(compression %zu usec, decompression %zu usec, bytes raw %zu, compressed %zu, savings ratio %0.2f%%)\n",
-                        name, compression_ut, decompression_ut,
+                        "(compression %llu usec, decompression %llu usec, bytes raw %zu, compressed %zu, savings ratio %0.2f%%)\n",
+                        name, (long long unsigned)compression_ut, (long long unsigned)decompression_ut,
                         bytes_uncompressed, bytes_compressed,
                         100.0 - (double)bytes_compressed * 100.0 / (double)bytes_uncompressed);
